# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a  server of IndustrialInternet.",
        "title": "IndustrialInternet项目API",
        "termsOfService": "https://www.topgoer.com",
        "contact": {
            "name": "www.topgoer.com",
            "url": "https://www.topgoer.com",
            "email": "example@some.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/login": {
            "get": {
                "description": "登录测试接口描述信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录"
                ],
                "summary": "登录测试接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\":200,\"msg\":\"\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/Response.ResponseStructure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Response.ResponseStructure": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}